add manual wtf_id parser; add human label checking
There's now an assert statement that says "everything you got labelled was
either noise, a dip, or a maybe_dip".

+++ src/utils/human_labelled_to_symlinks.py
@@ -13,6 +13,9 @@ def make_symlinks():
     df = pd.read_csv(cand_path, names=['kicid','human_label'], delimiter=', ',
             engine='python')
 
+    assert set(np.unique(df['human_label'])) == {'noise','dip','maybe_dip'}, \
+            'human_label should only be noise, dip, or maybe_dip'
+
     # Make symlinks for "dips", then for 'maybes"
     for id_type in ['dip', 'maybe_dip']:
         dipids = np.array(df[df['human_label']==id_type]['kicid'])
@@ -23,12 +26,18 @@ def make_symlinks():
             substr = 'dip' if id_type=='dip' else 'maybe'
             dstpath = '/home/luke/Dropbox/proj/cbp/results/real_search/'+\
                     'labelled_{:s}_symlinks/'.format(substr)+str(dipid)+'.png'
-            try:
-                os.symlink(srcpath, dstpath)
-            except:
-                print('symlink failed for {:s}'.format(str(dipid)))
+            if not os.path.exists(dstpath):
+                try:
+                    os.symlink(srcpath, dstpath)
+                except:
+                    print('symlink failed for {:s}'.format(str(dipid)))
 
         print('\nSymlinked human-labelled {:s}!'.format(id_type))
+
+    # Make a list of wtfs
+    wtfids = np.array(df[df['human_label']=='wtf']['kicid'])
+    np.savetxt('../../results/real_search/wtf_ids.txt', wtfids, fmt='%d')
+
 
 
 if __name__ == '__main__':
